requirements.txt
# Core Framework
fastapi>=0.108.0
uvicorn>=0.25.0
python-multipart>=0.0.6
pydantic>=2.5.3

# PDF Processing (Ingestion Agent)
pdfplumber>=0.10.3
PyMuPDF>=1.24.0
pytesseract>=0.3.10
python-docx>=1.1.0

# LLM Integration (Content Understanding & Explanation Agents)
groq>=0.4.1
python-dotenv>=1.0.0

# Image Processing (Slide Generation)
Pillow>=10.1.0
matplotlib>=3.8.2

# TTS (Text-to-Speech Agent)
edge-tts>=6.1.9

# Video Processing (Video Composition Agent)
moviepy<2.0.0
ffmpeg-python>=0.2.0
mutagen>=1.47.0  # For reading audio metadata

# Wav2Lip Dependencies (Avatar Agent)
# Note: Wav2Lip has its own requirements.txt, install separately
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.26.4
opencv-python>=4.8.1.78
librosa==0.9.1
scipy>=1.11.4
tqdm>=4.66.1

# Utilities
requests>=2.31.0
aiofiles>=23.2.1
playwright>=1.41.0
